# time series: Parallel Queries + Benchmark