--- a/pkg/ts/server.go
+++ b/pkg/ts/server.go
@@ -24,6 +24,7 @@ import (
 
 	"github.com/cockroachdb/cockroach/pkg/ts/tspb"
 	"github.com/cockroachdb/cockroach/pkg/util/log"
+	"github.com/cockroachdb/cockroach/pkg/util/stop"
 	gwruntime "github.com/grpc-ecosystem/grpc-gateway/runtime"
 )
 
@@ -31,21 +32,40 @@ const (
 	// URLPrefix is the prefix for all time series endpoints hosted by the
 	// server.
 	URLPrefix = "/ts/"
+	// queryWorkerMax is the default maximum number of worker goroutines that
+	// the time series server can use to service incoming queries.
+	queryWorkerMax = 250
 )
+
+// ServerConfig provides a means for tests to override settings in the time
+// series server.
+type ServerConfig struct {
+	// The maximum number of query workers used by the server. If this
+	// value is zero, a default non-zero value is used instead.
+	QueryWorkerMax int
+}
 
 // Server handles incoming external requests related to time series data.
 type Server struct {
 	log.AmbientContext
-	db *DB
+	db        *DB
+	stopper   *stop.Stopper
+	workerSem chan struct{}
 }
 
 // MakeServer instantiates a new Server which services requests with data from
 // the supplied DB.
-func MakeServer(ambient log.AmbientContext, db *DB) Server {
+func MakeServer(ambient log.AmbientContext, db *DB, cfg ServerConfig, stopper *stop.Stopper) Server {
 	ambient.AddLogTag("ts-srv", nil)
+	queryWorkerMax := queryWorkerMax
+	if cfg.QueryWorkerMax != 0 {
+		queryWorkerMax = cfg.QueryWorkerMax
+	}
 	return Server{
 		AmbientContext: ambient,
 		db:             db,
+		stopper:        stopper,
+		workerSem:      make(chan struct{}, queryWorkerMax),
 	}
 }
 
@@ -79,27 +99,79 @@ func (s *Server) Query(
 	}
 
 	response := tspb.TimeSeriesQueryResponse{
-		Results: make([]tspb.TimeSeriesQueryResponse_Result, 0, len(request.Queries)),
+		Results: make([]tspb.TimeSeriesQueryResponse_Result, len(request.Queries)),
 	}
-	for _, query := range request.Queries {
-		datapoints, sources, err := s.db.Query(
-			ctx,
-			query,
-			Resolution10s,
-			sampleNanos,
-			request.StartNanos,
-			request.EndNanos,
-		)
-		if err != nil {
-			return nil, grpc.Errorf(codes.Internal, err.Error())
-		}
-		result := tspb.TimeSeriesQueryResponse_Result{
-			Query:      query,
-			Datapoints: datapoints,
-		}
 
-		result.Sources = sources
-		response.Results = append(response.Results, result)
+	// Defer cancellation of context passed to worker tasks; if main task
+	// returns early, worker tasks should be torn down quickly.
+	ctx, cancel := context.WithCancel(ctx)
+	defer cancel()
+
+	// Channel which workers use to report their result, which is either an
+	// error or nil (when successful).
+	workerOutput := make(chan error)
+
+	// Start a task which is itself responsible for starting per-query worker
+	// tasks. This is needed because RunLimitedAsyncTask can block; in the
+	// case where a single request has more queries than the semaphore limit,
+	// a deadlock would occur because queries cannot complete until
+	// they have written their result to the "output" channel, which is
+	// processed later in the main function.
+	if err := s.stopper.RunAsyncTask(ctx, func(ctx context.Context) {
+		for queryIdx, query := range request.Queries {
+			queryIdx := queryIdx
+			query := query
+			if err := s.stopper.RunLimitedAsyncTask(
+				ctx,
+				s.workerSem,
+				true, /* wait */
+				func(ctx context.Context) {
+					datapoints, sources, err := s.db.Query(
+						ctx,
+						query,
+						Resolution10s,
+						sampleNanos,
+						request.StartNanos,
+						request.EndNanos,
+					)
+					if err == nil {
+						response.Results[queryIdx] = tspb.TimeSeriesQueryResponse_Result{
+							Query:      query,
+							Datapoints: datapoints,
+						}
+						response.Results[queryIdx].Sources = sources
+					}
+					select {
+					case workerOutput <- err:
+					case <-ctx.Done():
+					}
+				},
+			); err != nil {
+				// Stopper has been closed and is draining. Return an error and
+				// exit the worker-spawning loop.
+				select {
+				case workerOutput <- err:
+				case <-ctx.Done():
+				}
+				return
+			}
+		}
+	}); err != nil {
+		return nil, err
+	}
+
+	for range request.Queries {
+		select {
+		case err := <-workerOutput:
+			if err != nil {
+				// Return the first error encountered. This will cancel the
+				// worker context and cause all other in-progress workers to
+				// exit.
+				return nil, err
+			}
+		case <-ctx.Done():
+			return nil, ctx.Err()
+		}
 	}
 
 	return &response, nil