Allows the server to run the individual queries in an incoming
QueryRequest in parallel. This is a significant performance improvement over the
previous design, which executed the individual queries serially; typical UI
query requests routinely request data for more than 50-100 series at once.

Query worker count is managed via the stoppers "StartLimitedAsyncTask", which
limits the total count of workers for the entire server across all queries. The
limit has been arbitrarily set at 250, but can be adjusted for tests via a
TestingKnob (also added in this commit).

Also includes a commit to add a benchmark test for multi-query requests. This is added in a preceding commit, in order to demonstrate the difference in performance from the old serial execution. Benchstat output (50 runs):

```
name           old time/op  new time/op  delta
ServerQuery-8  9.76ms ± 9%  4.11ms ±13%  -57.91%  (p=0.000 n=50+50)
```

<!-- Reviewable:start -->
---

This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/10250)

<!-- Reviewable:end -->