@tschottdorf 2016-10-26
I'm curious - shouldn't parallel `DistSender` obviate parallel execution here?