@mrtracy 2016-10-26
No, what needs to be called in parallel is the `tsdb.Query()` method: it needs to be called once for each query in an incoming request.  The `Query()` method itself calls distsender to look up data; it is also responsible for performing all of our time series downsampling/aggregation/interpolation magic on the data it finds.