@petermattis 2016-10-26
> No, what needs to be called in parallel is the tsdb.Query() method: it needs to be called once for each query in an incoming request. 

Isn't that due to the interface to `tsdb.Query`? I could imagine a different API where `tsdb.Query` populated a `client.Batch` and a separate method processed the results from it. Certainly that would be a more significant refactoring than this PR.