@mrtracy 2016-10-26
That would be a considerable refactoring, and I don't think it would even be better; unless results from DistSender are streaming, we would have to complete the query of all data before we begin processing *any* data, whereas this method can begin processing data for each query as soon as data is available.