@tamird 2016-10-27
Reviewed 1 of 1 files at r1, 3 of 4 files at r2.
Review status: 3 of 4 files reviewed at latest revision, 10 unresolved discussions.

---

*[pkg/ts/server.go, line 43 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/10250#-KV2ZiEddiodN_63ieUE:-KV2ZiEddiodN_63ieUF:bj8nfi7) ([raw file](https://github.com/cockroachdb/cockroach/blob/3e9164c87770a0a43ea0e6aa284238b4e7807bd8/pkg/ts/server.go#L43)):*
> ```Go
> // TimeSeriesTestingKnobs provides a means for tests to override settings in the
> // test server that are not normally overridden.
> type TimeSeriesTestingKnobs struct {
> ```

I think this is getting ahead of yourself, and also it's not really appropriate for this to be a "testing knob". It's really more akin to a config object, like storage.StoreConfig.

For now, you can just pass an integer to MakeServer, and if that integer is zero, use a default.

---

*[pkg/ts/server.go, line 112 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/10250#-KV2bI_rZGbgv5cbyam5:-KV2Zz66CZoe4RfHMGX4:b-2ndw6t) ([raw file](https://github.com/cockroachdb/cockroach/blob/3e9164c87770a0a43ea0e6aa284238b4e7807bd8/pkg/ts/server.go#L112)):*
> ```Go
> 	// in the same order as the queries in the request.
> 	type queryResultWithError struct {
> 		queryNum int
> ```

s/Num/Idx/g

---

*[pkg/ts/server.go, line 129 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/10250#-KV2bjODJ1jTHTog3rzB:-KV2bjODJ1jTHTog3rzC:bj0p6df) ([raw file](https://github.com/cockroachdb/cockroach/blob/3e9164c87770a0a43ea0e6aa284238b4e7807bd8/pkg/ts/server.go#L129)):*
> ```Go
> 	// they have written their result to the "output" channel, which is
> 	// processed later in the main function.
> 	if err := s.stopper.RunAsyncTask(ctx, func(loopCtx context.Context) {
> ```

i think you actually do want to shadow `ctx` in all these closures; each closure should only be able to access the innermost context.

---

*[pkg/ts/server.go, line 130 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/10250#-KV2bKliXPf4WSVIsmnH:-KV2a6sx5kYiMQCLZsBP:b-y2oda3) ([raw file](https://github.com/cockroachdb/cockroach/blob/3e9164c87770a0a43ea0e6aa284238b4e7807bd8/pkg/ts/server.go#L130)):*
> ```Go
> 	// processed later in the main function.
> 	if err := s.stopper.RunAsyncTask(ctx, func(loopCtx context.Context) {
> 		for queryNum := 0; queryNum < len(request.Queries); queryNum++ {
> ```

`for queryIdx, query := range request.Queries` ?

---

*[pkg/ts/server.go, line 140 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/10250#-KV2asIuDpHcQPSVp4ka:-KV2asIvByOCh67Ypt-S:b-vel09i) ([raw file](https://github.com/cockroachdb/cockroach/blob/3e9164c87770a0a43ea0e6aa284238b4e7807bd8/pkg/ts/server.go#L140)):*
> ```Go
> 
> 			queryNum := queryNum
> 			if err := s.stopper.RunLimitedAsyncTask(
> ```

this can block for some time even after the context has been cancelled, can't it?

I think you should enhance `Run{,Limited}AsyncTask` to watch for cancellation on the context it's provided, which will allow you to omit the `select` above and will actually behave more as you intend.

---

*[pkg/ts/server.go, line 192 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/10250#-KV2agGqm9EuN92DspBP:-KV2agGqm9EuN92DspBQ:b-r0rtef) ([raw file](https://github.com/cockroachdb/cockroach/blob/3e9164c87770a0a43ea0e6aa284238b4e7807bd8/pkg/ts/server.go#L192)):*
> ```Go
> 	}
> 
> 	completed := 0
> ```

move into the loop

---

*[pkg/ts/server.go, line 203 at r2](https://reviewable.io:443/reviews/cockroachdb/cockroach/10250#-KV2ak25_Xal940D0BW2:-KV2ak25_Xal940D0BW3:bbxjote) ([raw file](https://github.com/cockroachdb/cockroach/blob/3e9164c87770a0a43ea0e6aa284238b4e7807bd8/pkg/ts/server.go#L203)):*
> ```Go
> 			}
> 			response.Results[queryResult.queryNum] = queryResult.result
> 			completed++
> ```

move into the `for` line

---

*[pkg/ts/server_test.go, line 265 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/10250#-KV2YbK6W-14qFFGg4LU:-KV2YbK7iQS9FnUwJOrE:b40lr9z) ([raw file](https://github.com/cockroachdb/cockroach/blob/dd5135d1e7584fc2ba14acc39d6f54cb75e868b1/pkg/ts/server_test.go#L265)):*
> ```Go
> 	b.ResetTimer()
> 
> 	queries := make([]tspb.Query, 0, seriesCount)
> ```

why include the construction of this slice in the benchmark time?

---

*[pkg/ts/server_test.go, line 275 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/10250#-KV2YgSN_LXK1XXXVwkU:-KV2YgSN_LXK1XXXVwkV:b321e9v) ([raw file](https://github.com/cockroachdb/cockroach/blob/dd5135d1e7584fc2ba14acc39d6f54cb75e868b1/pkg/ts/server_test.go#L275)):*
> ```Go
> 		if _, err := client.Query(context.Background(), &tspb.TimeSeriesQueryRequest{
> 			StartNanos: 0 * 1e9,
> 			EndNanos:   500 * 1e9,
> ```

math.MaxInt64? or whatever the correct type.

---

*[pkg/ts/server_test.go, line 291 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/10250#-KV2YrGRmVZMaxo5EQF4:-KV2YrGRmVZMaxo5EQF5:bisx3tb) ([raw file](https://github.com/cockroachdb/cockroach/blob/dd5135d1e7584fc2ba14acc39d6f54cb75e868b1/pkg/ts/server_test.go#L291)):*
> ```Go
> }
> 
> func populateSeries(t testing.TB, seriesCount, sourceCount int, tsdb *ts.DB) {
> ```

please return an error instead of Fataling inside.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/10250)*
<!-- Sent from Reviewable.io -->